Primeiro Teste API/requirements.txt
Flask==3.0.0
Flask-RESTful==0.3.10
SQLAlchemy==2.0.27
bcrypt==4.0.1
PyDrive==1.3.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1